Create common base class for DataGridViewModels
