+++ src/User Interface/ViewModels/BibliographyEditViewModel.cs
@@ -5,7 +5,7 @@ using DigitalProduction.Maui.ViewModels;
 
 namespace BibTeXManager.ViewModels;
 
-public partial class BibliographyEditViewModel : DataGridBaseViewModel<BibEntry>
+public partial class BibliographyEditViewModel : BibiographyPartDataGridBaseViewModel<BibEntry>
 {
 	#region Construction
 
@@ -25,8 +25,6 @@ public partial class BibliographyEditViewModel : DataGridBaseViewModel<BibEntry>
 	#endregion
 
 	#region Properties
-
-	public BibTeXProject	Project			{ get => BibTeXProject.Instance ?? throw new NullReferenceException("Project is null."); }
 
 	[ObservableProperty]
 	public partial bool		HasTemplates	{ get; set; } = false;
@@ -63,25 +61,7 @@ public partial class BibliographyEditViewModel : DataGridBaseViewModel<BibEntry>
 
 	#region File Menu
 
-	public async Task New()
-	{
-		if (Project.Bibliography != null)
-		{
-			Items = Project.Bibliography.Entries;
-		}
-	}
-
-	public async Task Open()
-	{
-		Items?.Clear();
-		Items = Project.Bibliography.Entries;
-	}
-
-	public async Task Close()
-	{
-		Items?.Clear();
-		Items = null;
-	}
+	protected override void AddItems() => Items = Project.Bibliography.Entries;
 
 	#endregion
